backend/requirements.txt
Django>=4.0
djangorestframework
pandas
scikit-learn
matplotlib
reportlab>=4.0.0
numpy
seaborn
plotly
xgboost
lightgbm
catboost
optuna
joblib
pickle-mixin
python-decouple
django-cors-headers
langgraph
langchain
langchain-google-genai
google-generativeai
openai
anthropic
